Fix issues with ad loading
Starting from Vungle SDK 3.2.0, delegate method
vungleSDKAdPlayableChanged is called when ad is ready to play, whereas
deprecated method vungleSDKhasCachedAdAvailable is called when ad is
loaded, but may not be ready to play yet.

The methods isAdPlayable and isCachedAdAvailable are the same, but the
second one is deprecated and will be removed soon.